backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.19
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
pydantic-settings==2.6.1
httpx==0.28.1
openai==1.57.0
python-dotenv==1.0.1
email-validator==2.2.0
bcrypt==4.0.1
greenlet
gunicorn
asyncpg
psycopg2-binary
google-generativeai>=0.8.3
groq
# OAuth
authlib
itsdangerous
# Cloudflare R2 (S3-compatible)
boto3
# Celery + Redis for background tasks
celery[redis]>=5.3.0
redis>=5.0.0

# Workflow Orchestration
langgraph>=0.0.10
langchain>=0.1.0

# Video/Audio Engine
moviepy>=1.0.3
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Payments
stripe>=8.0.0